code/requirements.txt
google-generativeai==0.8.3
sentence-transformers==3.4.1
scikit-learn==1.6.1
rank-bm25==0.2.2
pandas==2.2.3
python-dotenv==1.1.0
rich==14.0.0
numpy==2.2.5
